requirements.txt
annotated-types==0.7.0
anyio==4.7.0
astrapy==1.5.2
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
fastapi==0.115.6
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
jiter==0.8.2
openai==1.57.4
packaging==24.2
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
toml==0.10.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
uuid6==2024.7.10
uvicorn==0.34.0
